requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Django==4.2.6
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Faker==19.8.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.11.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
twilio==8.9.1
tzdata==2023.3
urllib3==2.0.6
yarl==1.9.2
